WIP refactor /edit route into component files
