+++ client/app/routes/playtrees.$playtree.edit.tsx
@@ -1,773 +1,20 @@
 import { LoaderFunctionArgs } from "@remix-run/node";
-import { useLoaderData, useLocation, useNavigate } from "@remix-run/react";
-import { Background, Controls, Handle, MarkerType, Position, ReactFlow, Node, NodeProps, EdgeProps, getBezierPath, Edge, BaseEdge, addEdge, OnConnect, useNodesState, useEdgesState, ConnectionLineComponent, EdgeLabelRenderer } from "@xyflow/react";
+import { useLoaderData } from "@remix-run/react";
+import { Background, Controls, MarkerType, ReactFlow, getBezierPath, addEdge, OnConnect, useNodesState, useEdgesState, ConnectionLineComponent } from "@xyflow/react";
 import '@xyflow/react/dist/style.css';
-import { useCallback, useEffect, useLayoutEffect, useMemo, useReducer, useState } from "react";
+import { useCallback, useEffect, useMemo, useReducer, useState } from "react";
 import invariant from "tiny-invariant";
-import NaturalNumberInputField from "~/components/NaturalNumberInputField"
-import SearchField, { queryString, SearchResult } from "~/components/SearchField";
-import { Content, jsonFromPlaytree, PlayEdge, PlayheadInfo, PlayNode, Playscope, Playtree, playtreeFromJson } from "../types";
-import React from "react";
+import { jsonFromPlaytree, PlayEdge, PlayNode, Playscope, Playtree, playtreeFromJson } from "../types";
 import Dagre from '@dagrejs/dagre';
-import { rgbToHex } from "@opentf/std";
+import { PlaytreeEditorAction, playtreeReducer } from "../reducers/playtree-editor";
+import PlaynodeComponent, { PlaynodeFlowData } from "../components/PlaynodeComponent";
+import PlayedgeComponent, { PlayedgeFlowData } from "../components/PlayedgeComponent";
+import { PlayConnectionLine } from "../components/PlayConnectionLine";
 
 export const loader = async ({ params }: LoaderFunctionArgs) => {
 	invariant(params.playtree)
 	const response = await fetch(`http://localhost:8080/playtrees/${params.playtree}`)
 	return await response.json()
-}
-
-type PlayheadProps = {
-	name: string;
-	nodeID: string;
-	dispatch: (action: PlaytreeEditorAction) => void;
-}
-
-function PlayheadComponent(props: PlayheadProps) {
-	const [name, setName] = useState(props.name)
-	const onNameChange = useCallback((evt: React.ChangeEvent<HTMLInputElement>) => {
-		setName(evt.target.value)
-		props.dispatch({
-			type: "updated_playhead",
-			nodeID: props.nodeID,
-			patch: {
-				name: evt.target.value,
-			}
-		})
-	}, [name])
-
-	const handleDeleteSelf = () => {
-		props.dispatch({
-			type: "deleted_playhead",
-			nodeID: props.nodeID
-		})
-	}
-
-	return (
-		<div id={name} className="group flex absolute -top-9 left-40 w-32">
-			<button onClick={handleDeleteSelf} className="bg-red-200 px-1 py-[2px] rounded-full text-xs absolute -top-3 -left-2 hidden group-hover:block">🗑️</button>
-			<div className="mr-2 bg-purple-300 px-2 py-1 rounded-md">💽</div>
-			<input value={name} onChange={onNameChange} className="bg-transparent w-full" />
-		</div>
-	)
-}
-
-type ContentProps = {
-	nodeID: string;
-	index: number;
-	color: string;
-	contentList: Content[];
-	onMoveUp?: () => void;
-	onMoveDown?: () => void;
-	onDeleteSelf: (index: number) => () => void;
-	onUpdateContentList: React.Dispatch<React.SetStateAction<Content[]>>;
-	dispatch: (action: PlaytreeEditorAction) => void;
-}
-
-function ContentComponent(props: ContentProps) {
-	const [mult, setMult] = useState<string>(props.contentList[props.index].mult.toString())
-	const [repeat, setRepeat] = useState<string>(props.contentList[props.index].repeat.toString())
-	const handleChangeMult = (event: React.ChangeEvent<HTMLInputElement>) => {
-		let inputAsNumber = Number(event.target.value)
-		if (event.target.value === "" || (Number.isInteger(inputAsNumber) && inputAsNumber >= 0)) {
-			if (event.target.value === "") {
-				inputAsNumber = 1
-			}
-			setMult(event.target.value)
-			const newContentList = [...props.contentList]
-			newContentList[props.index].mult = inputAsNumber
-			props.onUpdateContentList(newContentList)
-			props.dispatch({ type: "updated_playnode", nodeID: props.nodeID, patch: { content: newContentList } })
-		}
-	}
-	const handleChangeRepeat = (event: React.ChangeEvent<HTMLInputElement>) => {
-		let inputAsNumber = Number(event.target.value)
-		if (event.target.value === "" || event.target.value === "-" || (Number.isInteger(inputAsNumber) && inputAsNumber >= -1)) {
-			if (event.target.value === "") {
-				inputAsNumber = 1
-			} else if (event.target.value === "-") {
-				inputAsNumber = -1
-			}
-			setRepeat(event.target.value)
-			const newContentList = [...props.contentList]
-			newContentList[props.index].repeat = inputAsNumber
-			props.onUpdateContentList(newContentList)
-			props.dispatch({ type: "updated_playnode", nodeID: props.nodeID, patch: { content: newContentList } })
-		}
-	}
-	return (
-		<li key={props.contentList[props.index].id} className={`border border-${props.color}-600 bg-${props.color}-200 font-markazi flex`}>
-			{props.onMoveUp ? <button className="w-fit ml-1" title="Move Content Up In List" onClick={props.onMoveUp}>⬆️</button> : <div className="ml-5" />}
-			{props.onMoveDown ? <button className="w-fit ml-1" title="Move Content Down In List" onClick={props.onMoveDown}>⬇️</button> : <div className="ml-5" />}
-			<span className="w-full ml-3">{props.contentList[props.index].name}</span>
-			<input id="mult" name="mult" value={mult} onChange={handleChangeMult} className={`bg-${props.color}-200 w-6`} />
-			<input id="repeat" name="repeat" value={repeat} onChange={handleChangeRepeat} className={`bg-${props.color}-200 w-6`} />
-			<button className="w-fit mr-1" title="Delete Content" onClick={props.onDeleteSelf(props.index)}>❌</button>
-		</li>
-	)
-}
-
-export type PlayNodeFlow = Node<{
-	playnode: PlayNode;
-	playhead: PlayheadInfo | null;
-	scopes: Playscope[];
-	dispatch: (action: PlaytreeEditorAction) => void;
-	handleDeletePlaynode: (id: string) => void;
-}, 'play'>;
-
-function PlayNodeFlow(props: NodeProps<PlayNodeFlow>) {
-	const [adding, setAdding] = useState<boolean>(false)
-	const [scopeView, setScopeView] = useState<boolean>(false)
-
-	const [contentList, setContentList] = useState<Content[]>(props.data.playnode.content)
-
-	const highestID = props.data.playnode.content.map(content => parseInt(content.id)).reduce((id1, id2) => Math.max(id1, id2), -1)
-	const [contentID, setContentID] = useState<number>(highestID + 1)
-
-	const getNextID = useCallback(() => {
-		const nextID = contentID
-		setContentID(c => c + 1)
-		return nextID
-	}, [contentID])
-
-	const handleAddBegin = useCallback((_: any) => {
-		setAdding(true)
-	}, [])
-
-	const handleContentSelect = useCallback((newContent: SearchResult): boolean => {
-		if (newContent.uri === null) {
-			return false
-		}
-		const newContentList = structuredClone(contentList)
-		newContentList.push({ id: getNextID().toString(), type: "spotify-track", name: queryString(newContent), uri: newContent.uri, mult: 1, repeat: -1 })
-		setContentList(newContentList)
-		props.data.dispatch({ type: "updated_playnode", nodeID: props.data.playnode.id, patch: { content: newContentList } })
-		setAdding(false)
-		return false
-	}, [adding, contentList])
-
-	const handleSearchFocusOut = useCallback((event: FocusEvent) => {
-		setAdding(false)
-	}, [])
-
-	const handleChangeName = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
-		props.data.dispatch({ type: "updated_playnode", nodeID: props.data.playnode.id, patch: { name: event.target.value } })
-	}, [props.data.playnode.name]);
-
-	const handleChangeRepeat = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
-		let inputAsNumber = Number(event.target.value)
-		if (event.target.value === "" || event.target.value === "-" || (Number.isInteger(inputAsNumber) && inputAsNumber >= -1)) {
-			if (event.target.value === "") {
-				inputAsNumber = 1
-			}
-			if (event.target.value === "-") {
-				inputAsNumber = -1
-			}
-			if (props.data) {
-				
-			}
-		}
-	}, [props.data.playnode.repeat])
-
-	const handleTogglePlaynodeType = useCallback(() => {
-		const otherType: PlayNode["type"] = props.data.playnode.type === "sequence" ? "selector" : "sequence"
-		props.data.dispatch({ type: "updated_playnode", nodeID: props.data.playnode.id, patch: { type: otherType } })
-	}, [props.data.playnode.type])
-
-	const handleToggleScope = useCallback(() => {
-		setScopeView(sv => !sv)
-	}, [])
-
-	const handleMoveUp = useCallback((index: number) => () => {
-		if (index <= 0) {
-			return
-		}
-		const newContentList = structuredClone(contentList)
-		newContentList[index - 1] = contentList[index]
-		newContentList[index] = contentList[index - 1]
-		setContentList(newContentList)
-		props.data.dispatch({ type: "updated_playnode", nodeID: props.data.playnode.id, patch: { content: newContentList } })
-	}, [contentList])
-
-	const handleMoveDown = useCallback((index: number) => () => {
-		if (index + 1 >= contentList.length) {
-			return
-		}
-		const newContentList = structuredClone(contentList)
-		newContentList[index + 1] = contentList[index]
-		newContentList[index] = contentList[index + 1]
-		setContentList(newContentList)
-	}, [contentList])
-
-	const handleDeleteContent = useCallback((index: number) => () => {
-		const newContentList = structuredClone(contentList)
-		newContentList.splice(index, 1)
-		setContentList(newContentList)
-		props.data.dispatch({ type: "updated_playnode", nodeID: props.data.playnode.id, patch: { content: newContentList } })
-	}, [contentList])
-
-	const handleDeleteSelf = useCallback(() => {
-		props.data.handleDeletePlaynode(props.data.playnode.id)
-	}, [])
-
-	const isSequence = props.data.playnode.type === "sequence"
-	const color = isSequence ? "green" : "amber"
-
-	const handleDrop = (event: any) => {
-		event.preventDefault();
-		props.data.dispatch({ type: "added_playhead", nodeID: props.data.playnode.id })
-	}
-
-	const handleAddScope : React.FormEventHandler<HTMLFormElement> = event => {
-		event.preventDefault()
-		const form = event.target;
-		const formData = new FormData(form as any);
-
-		props.data.dispatch({type: "added_scope_to_playnode", index: parseInt((formData.get("scope-id") as string).toString()), nodeID: (formData.get("node-id") as string).toString() })
-	}
-
-	return (
-		<React.Fragment key={props.id}>
-			<div>{props.data.playhead ? <PlayheadComponent name={props.data.playhead.name} nodeID={props.id} dispatch={(x) => props.data.dispatch(x)} /> : null}</div>
-			<Handle type="target" isConnectableStart={false} position={Position.Top} style={{ width: 12, height: 12 }} />
-			{
-				props.selected ?
-					<div className={`border-${color}-600 bg-${color}-100 border-4 rounded-xl w-64 p-4 text-${color}-600`} onDrop={handleDrop} onDragOver={e => e.preventDefault()}>
-						<div className="mb-5">
-							<button className={`bg-${color}-300 rounded-lg px-2 py-1 absolute top-1 left-1`} onClick={handleTogglePlaynodeType} title={props.data.playnode.type}>{isSequence ? <>🔢</> : <>🎲</>}</button>
-							<button className={`bg-blue-300 rounded-lg px-2 py-1 absolute top-1 left-10`} onClick={handleToggleScope} title={scopeView ? "Toggle Song View" : "Toggle Scope View"}>{scopeView ? <>🎶</> : <>🔲</>}</button>
-							<button className={`bg-red-300 rounded-lg px-2 py-1 absolute top-1 right-1`} onClick={handleDeleteSelf} title="Delete Playnode">🗑️</button>
-						</div>
-						<input id="text" name="text" value={props.data.playnode.name} onChange={handleChangeName} className={`w-full bg-${color}-100 text-center`} />
-						<ul className="my-3">
-							{
-								scopeView ?
-								<>
-									{
-										props.data.playnode.scopes.map((scopeIndex, index) => {
-											const scope = props.data.scopes[scopeIndex]
-											return <li key={index} className={`font-markazi`} style={{color: scope.color}}>{scope.name}</li>
-										})
-									}
-									{
-										<form method="post" onSubmit={handleAddScope}>
-											<input type="hidden" name={"node-id"} value={props.data.playnode.id}></input>
-											<select name={"scope-id"} className="font-markazi">
-												{
-													props.data.scopes.map((scope, index) =>
-														props.data.playnode.scopes.includes(index) ? null
-														: <option key={index} value={index}>{scope.name}</option>
-													)
-												}
-											</select>
-											<button type="submit" className="border-black border-2 px-1 font-markazi">Add</button>
-										</form>
-									}
-								</>
-								 :
-								 <>
-								 	<div className="font-markazi">
-										<span className="mr-1">Repeat:</span>
-										<NaturalNumberInputField canBeInfinite={true} defaultValue={1} value={props.data.playnode.repeat} onChange={(n : number) => props.data.dispatch({ type: "updated_playnode", nodeID: props.data.playnode.id, patch: { repeat: n } })}/>
-									</div>
-									<div className="flex font-markazi"><div className="ml-14">Name</div><div className="ml-[3.75rem]">M</div><div className="ml-3">R</div></div>
-									{
-										contentList.map((content: Content, index: number) => {
-											return <ContentComponent key={content.id} nodeID={props.id} index={index} color={color} contentList={contentList}
-												onMoveUp={index > 0 ? handleMoveUp(index) : undefined}
-												onMoveDown={index + 1 < contentList.length ? handleMoveDown(index) : undefined}
-												onDeleteSelf={handleDeleteContent}
-												onUpdateContentList={setContentList}
-												dispatch={props.data.dispatch} />
-										})
-									}
-								 </>
-							}
-						</ul>
-						{
-							adding ?
-								<SearchField onContentSelect={handleContentSelect} onFocusOut={handleSearchFocusOut} /> :
-								<div className="flex"><button title="Add Content" className={`border-${color}-600 bg-${color}-400 border-2 rounded-full px-2 py-1 m-auto`} onClick={handleAddBegin}>➕</button></div>
-						}
-					</div> :
-					<div className={`border-${color}-600 bg-${color}-100 text-${color}-600 border-4 rounded-xl w-64 h-16 py-4 text-center`} onDrop={handleDrop} onDragOver={e => e.preventDefault()}>{props.data.playnode.name}</div>
-			}
-			<Handle type="source" position={Position.Bottom} id="a" style={{ width: 12, height: 12 }} />
-		</React.Fragment>
-	)
-}
-
-type PlayEdgeFlow = Edge<{
-	playedge: PlayEdge;
-	dispatch: (action: PlaytreeEditorAction) => void;
-}, 'play'>;
-
-function PlayEdgeFlow(props: EdgeProps<PlayEdgeFlow>) {
-	if (!props.data) {
-		return null
-	}
-
-	const initialRepeat = props.data.playedge.repeat ?? -1
-
-	const [repeatInputText, setRepeatInputText] = useState<string>(initialRepeat.toString())
-
-	const handleRepeatChange = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
-		let inputAsNumber = Number(event.target.value)
-		if (event.target.value === "" || event.target.value === "-" || (Number.isInteger(inputAsNumber) && inputAsNumber >= -1)) {
-			setRepeatInputText(event.target.value)
-			if (event.target.value === "") {
-				inputAsNumber = 1
-			}
-			if (event.target.value === "-") {
-				inputAsNumber = -1
-			}
-			props.data?.dispatch({
-				type: "updated_playedge", sourceID: props.source, targetID: props.target, patch: {
-					repeat: inputAsNumber
-				}
-			})
-		}
-	}, [repeatInputText])
-
-	const handleDeleteSelf = useCallback(() => {
-		props.data?.dispatch({type: "deleted_playedge", sourceID: props.source, targetID: props.target})
-	}, [])
-
-	const { sourceX, sourceY, targetX, targetY, markerEnd } = props;
-	let [edgePath, labelX, labelY] = getBezierPath(props)
-
-
-	if (sourceY > targetY) {
-		const distance = sourceY - targetY
-		const logDistance = Math.log(distance)
-		const scaledLogDistance = 40 * logDistance
-
-		const dx = sourceX - targetX
-		const nodeWidth = 500
-		const underAndLeft = dx < 0 && dx > -nodeWidth
-		const underAndRight = dx >= 0 && dx < nodeWidth
-
-		if (underAndLeft || underAndRight) {
-			const p0 = { x: sourceX, y: sourceY }
-			const p1 = { x: underAndRight ? sourceX - scaledLogDistance : sourceX + scaledLogDistance, y: sourceY + scaledLogDistance }
-			const p2 = { x: underAndRight ? targetX - scaledLogDistance : targetX + scaledLogDistance, y: targetY - scaledLogDistance }
-			const p3 = { x: targetX, y: targetY }
-
-			edgePath = `M ${p0.x} ${p0.y} C ${p1.x} ${p1.y} ${p2.x} ${p2.y} ${p3.x} ${p3.y}`
-
-			const t = (0.25)
-			const oneMinusT = 1 - t
-			const p0coeff = oneMinusT * oneMinusT * oneMinusT
-			const p1coeff = 3 * oneMinusT * oneMinusT * t
-			const p2coeff = 3 * oneMinusT * t * t
-			const p3coeff = t * t * t
-			labelX = p0coeff * p0.x + p1coeff * p1.x + p2coeff * p2.x + p3coeff * p3.x
-			labelY = p0coeff * p0.y + p1coeff * p1.y + p2coeff * p2.y + p3coeff * p3.y
-		}
-	}
-
-	return (
-		<React.Fragment key={props.id}>
-			<style>{
-				`.animate {
-                    stroke-dasharray: 10;
-                    animation: dash 0.5s linear;
-                    animation-iteration-count: infinite;
-                }
-
-                @keyframes dash {
-                    to {
-                        stroke-dashoffset: -20;
-                    }
-                }`
-			}
-			</style>
-			<BaseEdge path={edgePath} className={props.selected ? "animate" : ""} style={props.style} markerEnd={markerEnd} />
-			<EdgeLabelRenderer> {
-				props.selected ?
-					<div className="group bg-neutral-200 rounded-xl p-2 font-markazi" style={{
-						position: 'absolute',
-						transform: `translate(-50%, -50%) translate(${labelX}px,${labelY}px)`,
-						pointerEvents: 'all',
-						zIndex: 101
-					}}>
-						<div className="w-full h-fit flex content-evenly">
-							<div className="w-full"><button className="absolute -left-1 -top-1 hidden bg-red-300 text-xs rounded-full px-1 pt-1 group-hover:block" onClick={handleDeleteSelf}>🗑️</button></div>
-						</div>
-						<hr></hr>
-						<div className="w-24 flex">
-							<div className="w-full h-fit">Shares</div>
-							<div className="w-fit">|</div>
-							<NaturalNumberInputField onChange={n => {
-								props.data?.dispatch({ type: "updated_playedge", sourceID: props.source, targetID: props.target, patch: {shares: n}})
-							}} canBeInfinite={false} defaultValue={1} value={props.data.playedge.shares} />
-						</div>
-						<div className="w-24 flex">
-							<div className="w-full h-fit">Priority</div>
-							<div className="w-fit">|</div>
-							<NaturalNumberInputField onChange={n =>
-								props.data?.dispatch({type: "updated_playedge", sourceID: props.source, targetID: props.target, patch: {priority: n} })
-							} canBeInfinite={false} defaultValue={0} value={props.data.playedge.priority} />
-						</div>
-						<div className="w-24 flex">
-							<div className="w-full h-fit">Repeat</div>
-							<div className="w-fit">|</div>
-							<NaturalNumberInputField onChange={n =>
-								props.data?.dispatch({type: "updated_playedge", sourceID: props.source, targetID: props.target, patch: {repeat: n} })
-							} canBeInfinite={true} defaultValue={1} value={props.data.playedge.repeat} />
-						</div>
-					</div> :
-					<div className="bg-neutral-200 rounded-md px-1 font-markazi" style={{
-						position: 'absolute',
-						transform: `translate(-50%, -50%) translate(${labelX}px,${labelY}px)`,
-						pointerEvents: 'all'
-					}}>S={props.data.playedge.shares}, P={props.data.playedge.priority}, R={repeatInputText === "-1" ? "∞" : repeatInputText}</div>
-			}
-			</EdgeLabelRenderer>
-		</React.Fragment>
-	)
-}
-
-type LogMessage = {
-	type: "error" | "warning" | "success";
-	message: string;
-}
-
-type PlaytreeEditorState = {
-	playtree: Playtree,
-	unsavedChangesExist: boolean,
-	messageLog: LogMessage[],
-}
-
-type PlaytreeEditorAction = {
-	type: "loaded_playtree",
-	playtree: Playtree
-} | {
-	type: "added_playnode" | "saved_playtree" | "added_playscope",
-} | {
-	type: "updated_playnode",
-	nodeID: string,
-	patch: Partial<Omit<PlayNode, 'id' | 'next'>>
-} | {
-	type: "deleted_playnode",
-	nodeID: string
-} | {
-	type: "added_playedge" | "deleted_playedge",
-	sourceID: string,
-	targetID: string
-} | {
-	type: "updated_playedge",
-	sourceID: string,
-	targetID: string,
-	patch: Partial<Omit<PlayEdge, 'nodeID'>>
-} | {
-	type: "added_playhead",
-	nodeID: string
-} | {
-	type: "updated_playhead",
-	nodeID: string,
-	patch: Partial<Omit<PlayheadInfo, 'nodeID'>>
-} | {
-	type: "deleted_playhead",
-	nodeID: string,
-} | {
-	type: "logged_message",
-	message: LogMessage,
-} | {
-	type: "updated_playscope",
-	index: number,
-	patch: Partial<Playscope>
-} | {
-	type: "deleted_playscope",
-	index: number
-} | {
-	type: "added_scope_to_playnode",
-	index: number,
-	nodeID: string
-}
-
-const playtreeReducer = (state: PlaytreeEditorState, action: PlaytreeEditorAction): PlaytreeEditorState => {
-	const unsavedChangeOccurred = !["loaded_playtree", "saved_playtree", "logged_message"].includes(action.type)
-	switch (action.type) {
-		case "loaded_playtree": {
-			return {
-				...state,
-				playtree: action.playtree,
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "saved_playtree": {
-			return {
-				...state,
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "added_playnode": {
-			let maxValue = -1
-			state.playtree.nodes.forEach((_, id) => {
-				const x = parseInt(id)
-				if (maxValue < x) {
-					maxValue = x
-				}
-			})
-			const newPlaynode: PlayNode = {
-				id: (maxValue + 1).toString(),
-				name: "Playnode",
-				type: "sequence",
-				repeat: -1,
-				scopes: [],
-				content: [],
-				next: []
-			}
-			const newPlaynodes = structuredClone(state.playtree.nodes)
-			newPlaynodes.set(newPlaynode.id, newPlaynode)
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					nodes: newPlaynodes
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "updated_playnode": {
-			const newNodes = structuredClone(state.playtree.nodes)
-			let newPlaynode = newNodes.get(action.nodeID)
-			if (newPlaynode) {
-				newPlaynode = Object.assign(newPlaynode, action.patch)
-				newNodes.set(action.nodeID, newPlaynode)
-			}
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					nodes: newNodes
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "deleted_playnode": {
-			let newNodes = structuredClone(state.playtree.nodes)
-			newNodes.delete(action.nodeID)
-			Array.from(newNodes.values()).forEach(node => {
-				newNodes.set(node.id, {
-					...node,
-					next: node.next.filter(playedge => playedge.nodeID !== action.nodeID)
-				})
-			})
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					nodes: newNodes
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "added_playedge": {
-			const newNodes = structuredClone(state.playtree.nodes)
-			const sourceNode = newNodes.get(action.sourceID)
-
-			if (sourceNode) {
-				if (!sourceNode.next) {
-					sourceNode.next = []
-				}
-				sourceNode.next.push({
-					nodeID: action.targetID,
-					shares: 1,
-					priority: 0,
-					repeat: -1,
-				})
-				return {
-					...state,
-					playtree: {
-						...state.playtree,
-						nodes: newNodes
-					},
-					unsavedChangesExist: unsavedChangeOccurred,
-				}
-			}
-
-			return state
-		}
-		case "updated_playedge": {
-			const newNodes = structuredClone(state.playtree.nodes)
-			const sourceNode = newNodes.get(action.sourceID)
-
-			if (sourceNode && sourceNode.next) {
-				const playedgeIndex = sourceNode.next.findIndex(playedge => playedge.nodeID === action.targetID)
-				const playedge = sourceNode.next[playedgeIndex]
-				if (playedgeIndex !== -1) {
-					sourceNode.next.splice(playedgeIndex, 1, Object.assign(playedge, action.patch))
-					return {
-						...state,
-						playtree: {
-							...state.playtree,
-							nodes: newNodes
-						},
-						unsavedChangesExist: unsavedChangeOccurred
-					}
-				}
-			}
-			return state
-		}
-		case "deleted_playedge": {
-			const newNodes = structuredClone(state.playtree.nodes)
-			const sourceNode = newNodes.get(action.sourceID)
-
-			if (sourceNode && sourceNode.next) {
-				const playedgeIndex = sourceNode.next.findIndex(playedge => playedge.nodeID === action.targetID)
-				if (playedgeIndex !== -1) {
-					sourceNode.next.splice(playedgeIndex, 1)
-					return {
-						...state,
-						playtree: {
-							...state.playtree,
-							nodes: newNodes
-						},
-						unsavedChangesExist: unsavedChangeOccurred
-					}
-				}
-			}
-			return state
-		}
-		case "added_playhead": {
-			const newPlayroots = structuredClone(state.playtree.playroots)
-			const newPlayhead: PlayheadInfo = {
-				index: state.playtree.playroots.size,
-				name: "Playhead",
-			}
-			newPlayroots.set(action.nodeID, newPlayhead)
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					playroots: newPlayroots
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "updated_playhead": {
-			const newPlayroots = structuredClone(state.playtree.playroots)
-			const playroot = newPlayroots.get(action.nodeID)
-			if (playroot) {
-				const newPlayroot = Object.assign(playroot, action.patch)
-				newPlayroots.set(action.nodeID, newPlayroot)
-			}
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					playroots: newPlayroots
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "deleted_playhead": {
-			const newPlayroots = structuredClone(state.playtree.playroots)
-			newPlayroots.delete(action.nodeID)
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					playroots: newPlayroots
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "logged_message": {
-			const newMessageLog = [...state.messageLog]
-			newMessageLog.push(action.message)
-			return {
-				...state,
-				messageLog: newMessageLog,
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "added_playscope": {
-			const newPlayscopes = [...state.playtree.playscopes]
-			const index = newPlayscopes.length
-			const defaultColor = rgbToHex([255 * ((index & 4) >> 2), 255 * ((index & 2) >> 1), 255 * (index & 1)])
-			newPlayscopes.push({name: "Scope " + (index + 1), color: defaultColor})
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					playscopes: newPlayscopes
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "updated_playscope": {
-			const newPlayscopes = [...state.playtree.playscopes]
-			Object.assign(newPlayscopes[action.index], action.patch)
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					playscopes: newPlayscopes,
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "deleted_playscope": {
-			const newPlayscopes = [...state.playtree.playscopes]
-			newPlayscopes.splice(action.index, 1)
-			const newNodes = structuredClone(state.playtree.nodes)
-			newNodes.forEach(node => {
-				node.scopes = node.scopes.filter(scope => scope !== action.index).map(scope => {
-					return scope >= action.index ? scope + 1 : scope
-				})
-			})
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					playscopes: newPlayscopes,
-					nodes: newNodes
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-		case "added_scope_to_playnode": {
-			const nodesByScope = new Map<number, string>()
-			state.playtree.nodes.forEach(node => {
-				node.scopes.forEach(scopeID => {
-					if (!nodesByScope.has(scopeID)) {
-						nodesByScope.set(scopeID, node.id)
-					}
-				})
-			})
-
-			const newNodes = structuredClone(state.playtree.nodes)
-			newNodes.get(action.nodeID)?.scopes.push(action.index)
-
-			return {
-				...state,
-				playtree: {
-					...state.playtree,
-					nodes: newNodes
-				},
-				unsavedChangesExist: unsavedChangeOccurred
-			}
-		}
-	}
-}
-
-const PlayConnectionLine: ConnectionLineComponent<PlayNodeFlow> = ({ fromX, fromY, toX, toY }) => {
-	const [path] = getBezierPath({ sourceX: fromX, sourceY: fromY, targetX: toX, targetY: toY })
-	return (
-		<g>
-			<path
-				fill="none"
-				stroke="brown"
-				strokeWidth={2}
-				className="animated"
-				d={path}
-			>
-			</path>
-		</g>
-	)
 }
 
 type ScopeManagerProps = {
@@ -799,8 +46,8 @@ const ScopeManager = (props: ScopeManagerProps) => {
 }
 
 export default function PlaytreeEditor() {
-	const customFlowNodeTypes = useMemo(() => ({ play: PlayNodeFlow }), []);
-	const customFlowEdgeTypes = useMemo(() => ({ play: PlayEdgeFlow }), []);
+	const customFlowNodeTypes = useMemo(() => ({ play: PlaynodeComponent }), []);
+	const customFlowEdgeTypes = useMemo(() => ({ play: PlayedgeComponent }), []);
 
 	const initialPlaytree: Playtree | null = playtreeFromJson(useLoaderData())
 	if (initialPlaytree === null) {
@@ -823,7 +70,7 @@ export default function PlaytreeEditor() {
 		dispatch({ type: "deleted_playedge", sourceID: sourceID, targetID: targetID })
 	}
 
-	const initialFlownodes: PlayNodeFlow[] = Array.from(initialPlaytree.nodes.values()).map((playnode, index) => {
+	const initialFlownodeData: PlaynodeFlowData[] = Array.from(initialPlaytree.nodes.values()).map((playnode, index) => {
 		return {
 			key: playnode.id,
 			type: "play",
@@ -842,7 +89,7 @@ export default function PlaytreeEditor() {
 		}
 	})
 
-	const makeNewFlowedge = useCallback((playnode: PlayNode, playedge: PlayEdge): PlayEdgeFlow => {
+	const makeNewPlayedgeFlowData = useCallback((playnode: PlayNode, playedge: PlayEdge): PlayedgeFlowData => {
 		return {
 			id: playnode.id + "-" + playedge.nodeID,
 			type: "play",
@@ -864,17 +111,17 @@ export default function PlaytreeEditor() {
 		}
 	}, [])
 
-	let initialFlowedges: PlayEdgeFlow[] = []
+	let initialFlowedgeData: PlayedgeFlowData[] = []
 	initialPlaytree.nodes.forEach(playnode => {
 		if (playnode.next) {
 			playnode.next.forEach(playedge => {
-				initialFlowedges.push(makeNewFlowedge(playnode, playedge))
+				initialFlowedgeData.push(makeNewPlayedgeFlowData(playnode, playedge))
 			})
 		}
 	})
 
-	const [flownodes, setFlownodes, onFlownodesChange] = useNodesState<PlayNodeFlow>(initialFlownodes)
-	const [flowedges, setFlowedges, onFlowedgesChange] = useEdgesState<PlayEdgeFlow>(initialFlowedges)
+	const [flownodes, setFlownodes, onFlownodesChange] = useNodesState<PlaynodeFlowData>(initialFlownodeData)
+	const [flowedges, setFlowedges, onFlowedgesChange] = useEdgesState<PlayedgeFlowData>(initialFlowedgeData)
 
 	useEffect(() => {
 		const g = new Dagre.graphlib.Graph().setDefaultEdgeLabel(() => ({}));
@@ -906,12 +153,12 @@ export default function PlaytreeEditor() {
 		const sourcePlaynode = state.playtree.nodes.get(connection.source)
 		if (sourcePlaynode) {
 			const playedge = { nodeID: connection.target, shares: 1, priority: 0, repeat: -1 }
-			setFlowedges((eds) => addEdge(makeNewFlowedge(sourcePlaynode, playedge), eds))
+			setFlowedges((eds) => addEdge(makeNewPlayedgeFlowData(sourcePlaynode, playedge), eds))
 			dispatch({ type: "added_playedge", sourceID: connection.source, targetID: connection.target })
 		}
 	}, [state.playtree.nodes]);
 
-	const makeNewFlownode = useCallback((id: string) : PlayNodeFlow => {
+	const makeNewPlaynodeFlowData = useCallback((id: string) : PlaynodeFlowData => {
 		return {
 			id: id,
 			type: "play",
@@ -943,49 +190,49 @@ export default function PlaytreeEditor() {
 		// synchronize flownodes
 		setFlownodes(oldFlownodes => {
 			// next, upsert flownodes w/r/t the now-existing playnodes
-			const upsertedAndFilteredFlownodes : PlayNodeFlow[] = []
+			const upsertedAndFilteredPlaynodeFlowData : PlaynodeFlowData[] = []
 			// Note: this procedure is O(n^2) whose performance could be improved if necessary
 			state.playtree.nodes.forEach(playnode => {
-				let flownodeToUpsert = oldFlownodes.find(flownode => {
+				let playnodeFlowDataToUpsert = oldFlownodes.find(flownode => {
 					return flownode.id === playnode.id
-				}) ?? makeNewFlownode(playnode.id)
+				}) ?? makeNewPlaynodeFlowData(playnode.id)
 
-				flownodeToUpsert = {
-					...flownodeToUpsert,
+				playnodeFlowDataToUpsert = {
+					...playnodeFlowDataToUpsert,
 					data: {
-						...flownodeToUpsert.data,
+						...playnodeFlowDataToUpsert.data,
 						playnode: {...playnode},
 						scopes: state.playtree.playscopes,
 						playhead: state.playtree.playroots.get(playnode.id) ?? null
 					}
 				}
 
-				upsertedAndFilteredFlownodes.push(flownodeToUpsert)
+				upsertedAndFilteredPlaynodeFlowData.push(playnodeFlowDataToUpsert)
 			})
-			return upsertedAndFilteredFlownodes
+			return upsertedAndFilteredPlaynodeFlowData
 		})
 
 		// synchronize flowedges
 		setFlowedges(oldFlowedges => {
 			// upsert
-			const upsertedAndFilteredFlowedges : PlayEdgeFlow[] = []
+			const upsertedAndFilteredPlayedgeFlowData : PlayedgeFlowData[] = []
 			state.playtree.nodes.forEach(playnode => {
 				playnode.next?.forEach(playedge => {
-					const flowedgeToUpsert : PlayEdgeFlow = oldFlowedges.find(flowedge => {
+					const playedgeFlowDataToUpsert : PlayedgeFlowData = oldFlowedges.find(flowedge => {
 						return flowedge.id === `${playnode.id}-${playedge.nodeID}`
-					}) ?? makeNewFlowedge(playnode, playedge)
+					}) ?? makeNewPlayedgeFlowData(playnode, playedge)
 
-					if (flowedgeToUpsert.data) { // checking for typescript compiler
-						flowedgeToUpsert.data = {
-							...flowedgeToUpsert.data,
+					if (playedgeFlowDataToUpsert.data) { // checking for typescript compiler
+						playedgeFlowDataToUpsert.data = {
+							...playedgeFlowDataToUpsert.data,
 							playedge: playedge
 						}
 					}
 
-					upsertedAndFilteredFlowedges.push(flowedgeToUpsert)
+					upsertedAndFilteredPlayedgeFlowData.push(playedgeFlowDataToUpsert)
 				})
 			})
-			return upsertedAndFilteredFlowedges
+			return upsertedAndFilteredPlayedgeFlowData
 		})
 	}, [state.playtree])
